Replaced parameter used in SQLiteDatabase update
1. In VotingDatabase.java > updateInformation(...) method > sqLiteDatabase.update(VotingData.VotingInfo.TABLE_NAME, values, selection, args);

The "selection" parameter will result in failure of sqLiteDatabase.update as the syntax "parameter=null" is not permissible. Where arg[N] results in NULL (because the corresponding cursor.getString(N) = null), "parameter IS null" is expected instead.

Hence, under the assumption that EVENT_ID is unique, it will be used for updating our targeted row.

2. Minor update to registering interface activity_register.xml: Prevent user from entering newline in the username field.